requirements.txt
# Instagram API
instagrapi>=1.16.30

# Telegram Bot
python-telegram-bot==13.7

# Работа с изображениями
Pillow>=9.0.0

# База данных
SQLAlchemy>=1.4.0

# Многопоточность и асинхронность
# concurrent.futures - встроенный модуль, не требует установки
# asyncio - встроенный модуль, не требует установки

# Планировщик задач
schedule>=1.1.0
APScheduler>=3.6.3

# Прочие утилиты
python-dotenv>=0.19.0
requests>=2.26.0
loguru>=0.6.0

# Веб-драйверы (если используются для автоматизации)
selenium>=4.0.0
webdriver-manager>=3.5.0

# Обработка данных
pandas>=1.3.0

# Мониторинг системы
psutil>=5.9.0

# Работа с файлами
moviepy>=1.0.3

# Обработка видео для уникализации
opencv-python>=4.5.0
numpy>=1.21.0

# Работа с метаданными изображений
piexif>=1.1.3

selenium>=4.1.0
pytesseract>=0.3.8
Pillow>=9.0.0
webdriver-manager>=3.5.2

aiogram>=3.0.0
aiohttp>=3.8.0